finished wrapping everything into a class'
